[1/3]: Routing policies, add config in schema
This Series:
    Routing policies:

    Routing policy commands:
     lr-policy-add ROUTER PRIORITY MATCH ACTION [NEXTHOP]
          add a policy to router
     lr-policy-del ROUTER [PRIORITY [MATCH]]
     lr-policy-list ROUTER print policies for ROUTER

lr-policy-add : A policy is uniquely identified by priority and match.
Multiple policies can have the same priority.

lr-policy-del : takes priority and match as optional parameters. If priority is not specified, all the rules would be deleted. If priority is specified, all rules with that priority are deleted. If priority and match are specified, the policy with the given priority and match is deleted.

lr-policy-list : mandates ROUTER parameter.

This Patch:
 Add routing policies in the northbound schema.